WIP: move gdal related functions to destination modules
