+++ wradlib/georef/misc.py
@@ -14,7 +14,6 @@ Miscellaneous
    beam_height_n
    arc_distance_n
    get_earth_radius
-   get_shape_coordinates
 """
 
 import numpy as np
@@ -163,121 +162,3 @@ def get_earth_radius(latitude, sr=None):
                      (np.power(RADIUS_E, 2) * np.power(np.cos(latitude), 2) +
                       np.power(RADIUS_P, 2) * np.power(np.sin(latitude), 2)))
     return (radius)
-
-
-def get_shape_points(geom):
-    """
-    Extract coordinate points from given ogr geometry as generator object
-
-    If geometries are nested, function recurses.
-
-    .. versionadded:: 0.6.0
-
-    Parameters
-    ----------
-    geom : ogr.Geometry
-
-    Returns
-    -------
-    result : generator object
-        expands to Nx2 dimensional nested point arrays
-    """
-
-    type = geom.GetGeometryType()
-    if type:
-        # 1D Geometries, LINESTRINGS
-        if type == 2:
-            result = np.array(geom.GetPoints())
-            yield result
-        # RINGS, POLYGONS, MULTIPOLYGONS, MULTILINESTRINGS
-        elif type > 2:
-            # iterate over geometries and recurse
-            for item in geom:
-                for result in get_shape_points(item):
-                    yield result
-    else:
-        print("Unknown Geometry")
-
-
-def transform_geometry(geom, dest_srs):
-    """
-    Perform geotransformation to given destination SpatialReferenceSystem
-
-    It transforms coordinates to a given destination osr spatial reference
-    if a geotransform is neccessary.
-
-    .. versionadded:: 0.6.0
-
-    Parameters
-    ----------
-    geom : ogr.geometry
-    dest_srs: osr.SpatialReference
-        Destination Projection
-
-    Returns
-    -------
-    geom : ogr.Geometry
-        Transformed Geometry
-    """
-
-    # transform if not the same spatial reference system
-    if not geom.GetSpatialReference().IsSame(dest_srs):
-        geom.TransformTo(dest_srs)
-
-    return geom
-
-
-def get_shape_coordinates(layer, **kwargs):
-    """
-    Function iterates over gdal ogr layer features and packs extracted shape
-    coordinate points into nested ndarray
-
-    It transforms coordinates to a given destination osr spatial reference if
-    dest_srs is given and a geotransform is neccessary.
-
-    .. versionadded:: 0.6.0
-
-    Parameters
-    ----------
-    layer : ogr.Layer
-
-    Keywords
-    --------
-    dest_srs: osr.SpatialReference
-        Destination Projection
-    key : string
-        attribute key to extract from layer feature
-
-    Returns
-    -------
-    shp : nested :class:`numpy:numpy.ndarray`
-        Dimension of subarrays Nx2
-        extracted shape coordinate points
-    attrs : list
-        List of attributes extracted from features
-    """
-
-    shp = []
-
-    dest_srs = kwargs.get('dest_srs', None)
-    key = kwargs.get('key', None)
-    if key:
-        attrs = []
-    else:
-        attrs = None
-
-    for i in range(layer.GetFeatureCount()):
-        feature = layer.GetNextFeature()
-        if feature:
-            if key:
-                attrs.append(feature[key])
-            geom = feature.GetGeometryRef()
-            if dest_srs:
-                transform_geometry(geom, dest_srs)
-            # get list of xy-coordinates
-            reslist = list(get_shape_points(geom))
-            shp.append(np.squeeze(np.array(reslist)))
-
-    shp = np.squeeze(np.array(shp))
-
-    return shp, attrs
